+++ SDK-sources/OutbrainSDK/SmartFeed/SFHorizontalCell/SFHorizontalView/SFHorizontalView.h
@@ -9,12 +9,14 @@
 #import <UIKit/UIKit.h>
 #import "OBRecommendation.h"
 #import "OBSettings.h"
+#import "SFCollectionViewCell.h"
 
 @interface SFHorizontalView : UIView <UICollectionViewDataSource, UICollectionViewDelegate>
 
 typedef void(^OnRecommendationClick)(OBRecommendation *rec);
 typedef void(^OnAdChoicesIconClick)(NSURL *url);
 typedef CGSize(^CarouselItemSizeCallback)(void);
+typedef void(^ConfigureHorizontalItem)(SFCollectionViewCell * _Nonnull cell, OBRecommendation * _Nonnull rec);
 
 - (void) setupView;
 
@@ -25,7 +27,8 @@ typedef CGSize(^CarouselItemSizeCallback)(void);
 @property (nonatomic, strong) OBSettings * _Nullable settings;
 @property (nonatomic) OnRecommendationClick _Nonnull onRecommendationClick;
 @property (nonatomic) OnAdChoicesIconClick _Nonnull onAdChoicesIconClick;
-@property (nonatomic) CarouselItemSizeCallback _Nonnull carouselItemSizeCallback;
+@property (nonatomic) CarouselItemSizeCallback _Nullable carouselItemSizeCallback;
+@property (nonatomic) ConfigureHorizontalItem _Nullable configureHorizontalItem;
 
 // This section was private, now its here for the children classes
 @property (nonatomic, strong) UICollectionView *collectionView;