+++ SDK-sources/OutbrainSDK/SmartFeed/SFUtils/SFUtils.h
@@ -66,4 +66,7 @@
 
 +(NSURL *) appendParamsToVideoUrl:(OBRecommendationResponse *)response;
 
++(NSString *) getRecSourceText:(NSString *)recSource withSourceFormat:(NSString *)sourceFormat;
+
+
 @end