+++ SDK-sources/OutbrainSDK/SmartFeed/SmartFeedManager/SmartFeedManager.h
@@ -25,6 +25,8 @@
 
 -(CGSize) carouselItemSize;
 
+-(void) configureHorizontalItem:(SFCollectionViewCell * _Nonnull)sfCollectionViewCell withRec:(OBRecommendation * _Nonnull)rec;
+
 @end
 
 
@@ -47,6 +49,7 @@ typedef enum
     SFTypeStripVideoWithPaidRecAndTitle,
     SFTypeStripVideoWithPaidRecNoTitle,
     SFTypeGridTwoInRowWithVideo,
+    SFTypeBadType
 } SFItemType;
 
 @property (nonatomic, strong, readonly) NSString * _Nullable url;
@@ -93,14 +96,17 @@ typedef enum
     forItemAtIndexPath:(NSIndexPath * _Nonnull)indexPath;
 
 // Common Methods
--(NSString * _Nullable) sfItemTypeFor:(NSIndexPath *)indexPath;
+-(NSString * _Nullable) sfItemTypeStringFor:(NSIndexPath *)indexPath;
+
+-(SFItemType) sfItemTypeFor:(NSIndexPath *)indexPath;
 
 - (void) registerNib:(UINib * _Nonnull )nib withReuseIdentifier:( NSString * _Nonnull )identifier forWidgetId:(NSString *)widgetId;
 
 - (void) registerNib:(UINib * _Nonnull )nib withReuseIdentifier:( NSString * _Nonnull )identifier forSFItemType:(SFItemType)itemType;
 
-- (void) registerHeaderNib: (UINib * _Nonnull)nib withReuseIdentifier:( NSString * _Nonnull )identifier;
-
 - (void) setTransparentBackground: (BOOL)isTransparentBackground;
+
+-(NSArray * _Nullable) recommendationsForIndexPath:(NSIndexPath * _Nonnull)indexPath;
+
 
 @end