+++ SDK-sources/OutbrainSDK/SupportingFiles/OutbrainSDK.h
@@ -40,5 +40,6 @@ FOUNDATION_EXPORT const unsigned char DynamicOutbrainSDKVersionString[];
 #import <OutbrainSDK/SFHorizontalCollectionViewCell.h>
 #import <OutbrainSDK/SFHorizontalView.h>
 #import <OutbrainSDK/OBVideoWidget.h>
+#import <OutbrainSDK/OBUtils.h>